Fixed clip property's floating point round error. Added exit button.

+++ engine/source/http_trickplay_api_support.cpp
@@ -128,10 +128,10 @@ static void dump_ui_actors( ClutterActor * actor, JSON::Object * object )
 	clutter_actor_get_clip( actor, &cxoff, &cyoff, &cw, &ch );
 	Object clip;
 		// floats returned are very small nonzero numbers
-	clip["x"] = cxoff;
-	clip["y"] = cyoff;
-	clip["w"] = cw;
-	clip["h"] = ch;
+	clip["x"] = (long long)cxoff;
+	clip["y"] = (long long)cyoff;
+	clip["w"] = (long long)cw;
+	clip["h"] = (long long)ch;
 
     if ( !extra.empty() )
     {
